Updating related contents in other tabs won't refresh Preview panel and Page Editor #623
