+++ src/main/resources/assets/admin/common/js/content/resource/_module.ts
@@ -54,4 +54,3 @@
 ///<reference path='HasUnpublishedChildrenRequest.ts' />
 ///<reference path='ContentTreeSelectorItem.ts' />
 ///<reference path='ContentAndStatusTreeSelectorItem.ts' />
-///<reference path='ContentQueryRequestHelper.ts' />
